+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Remove `model` and `models` for input and output identifiers in tests. Replace by `shared` instead. ([#261](https://github.com/Substra/substra-tests/pull/261))
+- Moved `wait_task` and `wait_compute_task` from `substratest.client.Client` to `substra.Client` ([#263](https://github.com/Substra/substra-tests/pull/263))
 
 ## [0.42.0] - 2023-06-27
 
@@ -17,7 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Define test client (`substratest.client.Client`) as child class of `substra.Client` (#205)
 ([#257](https://github.com/Substra/substra-tests/pull/257))
-- Moved `wait_task` and `wait_compute_task` from `substratest.client.Client` to `substra.Client` ([#263](https://github.com/Substra/substra-tests/pull/263))
 
 ## [0.41.0] - 2023-06-12